small changes based on PR review comments
